Add tools to create test users with unexpired passwords
create-test-data.py comes from the freeipa-tools repo. It is
designed to create complex test data including nested groups,
hbac and sudo rules. It provides way more than we need here but
we may be able to build on it.

As such, this overrides some of its methods to generate a
number of users per host with a given username format.

Whether we can mangle this into creating the other entry types in
the future is TBD but I think it is possible.

The password is 'password'.

Users are created in the form: user#client#.<domain>

Signed-off-by: Rob Crittenden <rcritten@redhat.com>