A simple multi-threaded PAM tool to authenticate users
pam_loginuid.so needs to be disabled otherwise the session
creation will fail. This is because the thread id is different
from the pid so the kernel will reject a change to
/proc/self/loginuid

Usage is: ./thread --threads <num>

The default pam service is login. This can be overridden
with the --service option.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>