Refactor the singular test script into a plugin-based system
This adds a plugin framework based on the ipa-healthcheck
framework which makes this system a python package.

The core of the tests are broken into discrete methods that can
be overridden by individual tests as needed. It is likely that
this will need further tweaking.

Break each test out into its own plugin.

All arguments are passed around in a Click context.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>