Add test to simulate client authentications using a pam tool
The purpose of the test is to determine the maximum number of
authentications an IPA server is capable of. The general rule
of thumb has been 2-3000 clients per server. The intention of
this test is to back that up with numbers.

This is an IPA-centric test so only using IPA users and only
authenticating via SSSD using the PAM login service. A single
authentication for each user is done.

The basic idea is:

- Deploy some client machines, trying to gate their installation
  so that all are successful
- Create a bunch of users in IPA for each specific host in the form
  of user000client000.ipa.test
- Set Kerberos and LDAP passwords for every user to 'password'
- Run the pamtest tool with a number of threads and collect the
  result
- Pull the logs and determine the success rate

This should scale both vertically (number of threads) and
horizontally (more clients) so hopefully we can get a reproducible
and relevant number.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>

With many contributions from Antonio Torres <antorres@redhat.com>