APITest: update client generation settings
APITest was generating "amount" clients. However in this test we want
one client per 25 commands. Also increase available memory to 2GB since
in this test we need more memory for handling all the commands and we
don't need to deploy as many clients as in other tests such as
EnrollmentTest.

Signed-off-by: Antonio Torres <antorres@redhat.com>