AuthenticationTest: Add Active Directory support
Add Active Directory support to AuthenticationTest. This will create a
Windows Server host and deploy Active Directory server on it. After
that, test users are created into AD and trust is established with the
IPA server.

The pamtest tool is also updated to support launching threads for
authenticating Active Directory users.

Signed-off-by: Antonio Torres <antorres@redhat.com>