APITest: reduce number of SSH connections during test
APITest is optimized by reducing the number of SSH connections performed
during the test. To do so, we send all commands for a client in a single
connection, instead of opening a new one for each command. We also
reduce the connections by parsing and analyzing the results after we
fetch the log files, instead of doing it before.

Signed-off-by: Antonio Torres <antorres@redhat.com>