Add option to disable SELinux and enable forking in authtest
SSSD currently has a mutex when calling pam in threaded
more so it is effectively serialized. Add an option to call
PAM using fork() instead to run the authentications
simultaneously.

There is a lock in libsemanage which can cause failures
so disable the SELinux provicer as a workaround for that.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>