Increase number of maximum replicas
Instead of just allowing two replicas, allow up to 64. The topology for
these is built as a tree, with a maximum of 4 agreements per server.
Replicas need to be deployed in tiers, since some of them depend on
other replicas. Because of this, 4 tiers are predefined in the Ansible
playbook template. This also streamlines current playbooks, reducing the
number of tasks and number of playbooks created.

Signed-off-by: Antonio Torres <antorres@redhat.com>