Add options to check test results
Add options to check the test results, specifically the execution time,
both in total and per unit. The expected result can be passed when
executing the test, and it will produce an error when the test takes
longer than expected. This is useful for running the tests in a CI
environment, as the test will fail without needing a manual check.

Signed-off-by: Antonio Torres <antorres@redhat.com>