Remove --no-member, fix up GroupSizeTest tarballs
Using --no-member causes the membership to not be calculated
in IPA which makes the data return much faster. This is not
enabled in the UI or CLI by default. This change was left in
accidentally. It does improve the times a bit.

A syntax error in setting the GroupSizeTest archive name
resulted in no failure but there was no resulting tarball for
the execution.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>