Add script to generate certificates
This will be used for load, performance and migration testing.

It works by issuing a cert for a new service over and over
again. The certificate itself is dropped from the filesystem and
the service entry because there should be no need to have
potentially thousands of unique certificates lying around.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>